change Modal position for IngredientDetails
